v.1.6.6 beta initial

+++ build/powershell/benchmark.ps1
@@ -19,38 +19,36 @@ param(
 )
 Set-Location (Split-Path (Split-Path (Split-Path $script:MyInvocation.MyCommand.Path)))
 Write-Host "Checking For $Name Bechmarks"
-if($Platform -eq "windows"){"Removed Hashrate files" | Out-File ".\build\txt\benchcom.txt"}
-if(Test-Path ".\stats\*$($Name)_hashrate.txt*"){Remove-Item ".\stats\*$($Name)_hashrate.txt*" -Force}
-if(Test-Path ".\stats\*$($Name)_power.txt*"){Remove-Item ".\stats\*$($Name)_power.txt*" -Force}
-if(Test-Path ".\backup\*$($Name)_hashrate.txt*"){Remove-Item ".\backup\*$($Name)_hashrate.txt*" -Force}
-if(Test-Path ".\backup\*$($Name)_power.txt*"){Remove-Item ".\backup\*$($Name)_power.txt*" -Force}
-if(Test-Path ".\timeout\pool_block\pool_block.txt")
+
+Switch($Name)
  {
-  $NewPoolBlock = @()
-  $GetPoolBlock = Get-Content ".\timeout\pool_block\pool_block.txt" | ConvertFrom-Json
-  $GetPoolBlock | foreach {
-  if($($_.Coins) -ne $Name){$NewPoolBlock += $_}
-  else{Write-Host "Found $($_.Algo) in Pool Block file"}
-  }  
-  if($NewPoolBlock){$NewPoolBlock | ConvertTo-Json | Set-Content ".\timeout\pool_block\pool_block.txt"}
-  else{"" | Set-Content ".\timeout\pool_block\pool_block.txt"}
+  "timeout"
+   {
+    if(Test-Path ".\timeout"){Remove-Item ".\timeout" -Recurse -Force}
+    Write-Host "Removed All Timeouts" -ForegroundColor Green
+    if($Platform -eq "windows"){"Removed All Timeouts" | Out-File ".\build\txt\benchcom.txt"}
+   }
+   default
+   {
+    if(Test-Path ".\stats\*$($Name)_hashrate.txt*"){Remove-Item ".\stats\*$($Name)_hashrate.txt*" -Force}
+    if(Test-Path ".\stats\*$($Name)_power.txt*"){Remove-Item ".\stats\*$($Name)_power.txt*" -Force}
+    if(Test-Path ".\backup\*$($Name)_hashrate.txt*"){Remove-Item ".\backup\*$($Name)_hashrate.txt*" -Force}
+    if(Test-Path ".\backup\*$($Name)_power.txt*"){Remove-Item ".\backup\*$($Name)_power.txt*" -Force}
+    if(Test-Path ".\timeout\pool_block\pool_block.txt")
+    {
+     $NewPoolBlock = @()
+     $GetPoolBlock = Get-Content ".\timeout\pool_block\pool_block.txt" | ConvertFrom-Json
+     $GetPoolBlock | foreach {if($_.Algo -ne $Name){$NewPoolBlock += $_}else{Write-Host "Found $($_.Algo) in Pool Block file"}}
+     if($NewPoolBlock){$NewPoolBlock | ConvertTo-Json | Set-Content ".\timeout\pool_block\pool_block.txt"}
+    }
+    if(Test-Path ".\timeout\algo_block\algo_block.txt")
+    {
+     $NewPoolBlock = @()
+     $GetPoolBlock = Get-Content ".\timeout\algo_block\algo_block.txt" | ConvertFrom-Json
+     $GetPoolBlock | foreach {if($_.Algo -ne $Name){$NewPoolBlock += $_}else{Write-Host "Found $($_.Algo) in Pool Block file"}}
+     if($NewPoolBlock){$NewPoolBlock | ConvertTo-Json | Set-Content ".\timeout\pool_block\pool_block.txt"}
+    }
+    Write-Host "Removed all $Name stats and bans." -ForegroundColor Green
+    "Removed all $Name stats and bans." | Out-File ".\build\txt\benchcom.txt"
+   }
  }
- if(Test-Path ".\timeout\algo_block\algo_block.txt")
- {
-  $NewAlgoBlock = @()
-  $GetAlgoBlock = Get-Content ".\timeout\algo_block\algo_block.txt" | ConvertFrom-Json
-  $GetAlgoBlock | foreach {
-  if($($_.Coins) -ne $Name){$NewAlgoBlock += $_}
-  else{Write-Host "Found $($_.Algo) in Algo Block file"}
-  }
-  if($NewAlgoBlock){$NewAlgoBlock | ConvertTo-Json | Set-Content ".\timeout\algo_block\algo_block.txt"}
-  else{"" | Set-Content ".\timeout\algo_block\algo_block.txt"}
- if($Name -eq "timeout")
-  {
-   Remove-Item ".\timeout" -Recurse -Force
-   Write-Host "Removed All Timeouts"
-   if($Platform -eq "windows"){"Removed All Timeouts" | Out-File ".\build\txt\benchcom.txt"}
-  }
- }
-Write-Host "Removed Hashrate files"
-if($Platform -eq "windows"){"Removed Hashrate files" | Out-File ".\build\txt\benchcom.txt"}
